vh: sched: Remove adaptive CPU driver
The adaptive CPU driver registers a sched_switch tracepoint in order to
export statistics to userspace, which is an extremely horrible use of
sched_switch on a production device. The overhead of sched_switch is never
worth it in a production environment.

Remove the adaptive CPU feature entirely for its tracepoint transgressions.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Aarqw12 <lcockx@protonmail.com>

# Conflicts:
#	drivers/soc/google/vh/kernel/sched/Kbuild
#	drivers/soc/google/vh/kernel/sched/init.c